fix: 블루프린트에 사용될 수 없던 매개변수 타입 사용하고 있던 에러 수정

+++ SpikeBeach/Source/SpikeBeach/SpikeBeachGameInstance.cpp
@@ -174,7 +174,7 @@ void USpikeBeachGameInstance::SendRoomLeaveRequest()
 	WebSocket->Send(dataToSend.GetData(), dataToSend.Num(), true);
 }
 
-void USpikeBeachGameInstance::SendRoomReadyRequest(int16 team)
+void USpikeBeachGameInstance::SendRoomReadyRequest(int team)
 {
 	ReadyRequest room_ready_request;
 	room_ready_request.team = static_cast<Team>(team);
